Fix bug with HTTP 304 handling.
kdesrc-build added code a few commits ago to use LWP's mirror feature
for servers that support If-Modified-Since, but the main use-case
(kde_projects.xml) suffered a misfeature in the generator causing the
XML file to change approx. every five minutes anyways, which meant the
download always happened.

The server now writes the XML in a canonical order which means
If-Modified-Since is useful and works... and revealed that LWP's mirror
feature doesn't set the "is_success" flag (which really just means any
HTTP 2xx code). So check for 304 specifically and keep rolling...