Fix uninit warning for $repository.
The new Bzr support exposed that getOption(foo, 'module') will return undef if
the module doesn't have an option foo (instead of '' like normal). This causes
the test suite to give some warnings that aren't seen in normal code execution.

To fix we just use an empty string as the repository if getOption() returns
undef.