Fully name internal function to use with log_command.
log_command supports calling an internal subroutine instead of
exec()'ing a new process in the logged child. The subroutine is
identified by name.

With the move of log_command to ksb::Util, log_command now assumes that
the named function will be found in ksb::Util instead of the main
package, but none of the users of this feature were ported over.

This led to bug 297158 (kdesrc-build fails to --refresh-build a module)
since the 'prune_under_directory' routine used to refresh the build
directory wasn't being called by log_command anymore.

Thanks to Kevin Lyles for the prompt and informative report!

I've added a (very simple) test case to ensure the subroutines can be
called at all (a test which now passes).

BUG:297158
FIXED-IN:1.15