Refactor function export code.
In the act of removing a now-extraneous method prototype in the next
patch I noticed that I had not fixed ksb::Debug's import method to work
with method that had no prototype.

"Don't Repeat Yourself" being a good idea and all, I refactored both
methods to a common one that works on everything.