Support defaulting to stable branches from proj.k.o.
Valery Yundin has kindly submitted a patch (including documentation!)
resurrecting the old "use-stable-kde" option to allow for defaulting
kde-projects modules to the stable branch (if defined for that module).

This should track with changes to the stable branch as defined, so it
would be a good way to track the "not quite bleeding-edge" of
development.

A hearty thanks to Valery for making up a patch and then adapting it to
make use of the old use-stable-kde option.

BUG:297357
FIXED-IN:1.15