Depend on parent project modules if also in build.
This change allows a KDE Project module to depend on a KDE Project
module that is its immediate parent. For example,
kde/kdelibs/nepomuk-core would depend on kde/kdelibs if both are present
in the build, to ensure that kdelibs is cloned before nepomuk-core. To
be fully, absolutely correct we should probably depend on /any/ parent
present in the build, not just the immediate one. Maybe for later...

This makes existing dependency-data for kde-baseapps/{,konsole,kate}
redundant although I will leave it there for the benefit of other
projects that might be using it.