Support parsing dependency-data rev 2.
Torgny Nyblom has been working on a revision of the dependency-data
format to support using that information for
Jenkins/continuous-integration testing.

Right now we don't do much with it (perhaps later we can ensure that
branches are correct when building to keep from avoidable errors), but
this change should support reading in the new format.

Older kdesrc-build versions will throw an exception but continue on
without dependency-reordering and hopefully without further error.

CCMAIL:nyblom@kde.org