kdesrc-build-setup: Fix directory selector.
It was difficult to see the error message until I thought to remove the
'clearScreen' call before throwing the "dialog canceled" exception, but
the cause of bug 300911 (Setup quits unexpectedly when choosing a custom
install directory) is due to the dialog(1) command failing when it
couldn't create a window big enough for the directory selector.

Apparently the "height" parameter does not include all the curses
window decoration.

(Fixed-in flag reflects the kdesrc-build version)

BUG:300911
FIXED-IN:1.16