Add custom build command option.
Unfortunately I was not able to split out moving safe_make and
run_make_command with the actual implementation of this option
after-the-fact. Whoops.

This should implement support for a custom build command (and custom
install command, assuming the install can be performed by adding
'install' to the command line).

Hopefully that will be sufficient for 'ninja' or other buildsystems.
Note that this only changes the *build* command, the other parts of the
build system still needs to be supported in kdesrc-build (for now,
custom build systems probably wouldn't be too hard to implement if
necessary).

Assuming your proposed command is available in PATH and can be called
like make, doing something like this should suffice:

module (or module-set) foo
    custom-build-command ninja
end module (or module-set)

BUG:303692