Ensure STATUS_FILE always a valid filehandle.
There was a bug with printing to STATUS_FILE that wasn't guarded by
checking to see if $outfile was defined, which was causing errors when
printing that a module failed to update/build in --pretend mode.

I could have fixed it by adding that check but it's probably better to
have it always valid.

We can't set $outfile to '/dev/null' unconditionally since we decide
whether to create log/latest symlinks based on whether $outfile is set
or not. So instead I use a defined-or (//) operator in the open call.