Audit and correct list of config file options.
This audit was just comparing the list of config file options between
the documentation's list of options and the list contained in
kdesrc-build.

This resulted in adding an option (persistent-data-file) to kdesrc-build
to indicate that it's expected, and documenting an option
(use-clean-install) that was present but undocumented.

Original commit: 5785d4e8
https://invent.kde.org/sdk/kdesrc-build/-/commit/5785d4e8295cae284d0c1a710e46b85747115f5c