modularize: Move ksb::PhaseList to separate file.
This PhaseList was one of the first modules added as part of the
refactoring effort, but I don't think it has ended up very useful (or at
least not as I had designed).

This move also required fixing ksb::BuildContext which had not 'use'd
ksb::PhaseList before, it had worked by accident.