xsession: Disable session driver installation by default.
This may yet be moved to the kdesrc-build-setup script instead so
disable this by default so that those tracking kdesrc-build git don't
have to necessarily deal with this handling. It can still be manually
enabled.

This will cause a conflict with the 'modularize' branch, which should be
resolved by merging this change into the ksb::BuildContext Perl module.