git: Fix pretend mode branchname checks.
This should fix the big ugly error messages for this particular
scenario. The code in question is trying to determine a good branch name
that it can use, which means it need to verify that the requested branch
name *doesn't* exist. This would correspond to an exit code of 1 for the
given git command.

In pretend mode all safe_system() calls give an exit code of 0, so it
appears that there are no safe unused branch names.

Since even in pretend mode the source directory should exist in this
code path I switched to use a system() call (with the correct full
branch name to boot).

I'm not sure the other part of the bug is fixed (i.e. checking out a
local branch immediately after cloning or whatever dfaure was seeing) so
I'll leave open for now until I can do further testing.

CCBUG:311758