kde-projects: Really don't pull in modules w/out right branch.
Commit 63e21e9e1be32d9058d7d76aa731bead985f5838 filtered only based on a
module-set branch. This causes problems with global branches that might
have the same issue so we'll just filter out all such modules.

We probably need to use the scm()'s getBranch() method to have a 1:1
check in the presence of use-stable-kde, but even that should work due
to the way branch handling is performed with use-stable-kde (which
should always pick existing branches).

BUG:307694