kde-projects: Sanity check to avoid incl. all modules.
The last commit's first attempt had an error where the kde/foo/* syntax
would be stripped to '*', which the KDEXMLReader actually understands
perfectly. Unfortunately a bare '*' reads in all modules, including
websites and other unpleasant things.

I only caught this by chance, so add a sanity check to make this an
error in case I miss it in the future.