build-metadata: Handle ignored modules in pretend mode.
For whatever dumb reason I coded the "ignored kde-projects modules"
handling (kde-build-metadata/build-script-ignore) by having the code
that reads in the list of ignored modules go and immediately update the
build context.

For obvious reasons this only works when the update is actually
performed, which means ignore information wasn't used in pretend mode.

We fix this by moving the build context update to the rest of the
metadata handling.