kde-projects: Support ignoring given implicit modules.
It's kind of annoying to be unable to build all modules in a certain
kde-projects module grouping, *except* for some certain chosen ones.

Now you can: You can add the ignore-modules option in your module-sets
using kde-projects repository.

It accepts a space-separated list of module names (either kdefoo or
kde/module/kdefoo style) to ignore.

Although the intention is to only filter out matching modules in *that*
module-set, it would actually perform the filtering even if you
accidentally put it in the wrong module set due to a fluke of the
implementation (I recommend not relying on this, however).

The documentation has also been updated, including an example of the
usage.