Now forward persistent option changes to build proc as well.
Though this only works for Git modules, it should at least fix the
annoying "OMG YOUR REPO MOVED" warnings that never get fixed when in
async mode (which is the default).

Previously you simply never saw the warning due to the now-fixed output
squelching, now we need to fix this for good as well.