kde-projects: Fixup module name normalization.
Commit aa1e4b954279eba1c14d9b0bfa82f95a37992ed6 introduced changes to
how "proj" modules are first parsed (simply keeping the last path
component as entered by the user as the module name).

The unintended side-effect of this is that the user doesn't always have
to enter path ending in an actual repo. E.g. extragear/utils is not a
repo, it would get reduced to "utils", and later KDEXMLReader would
effectively look for all repos matching "*/utils/*".

For more fun, do what I did and try "kdegraphics/libs" => "libs" =>
"*/libs/*"...

Anyways, the fix here is to use the normalized module name as the module
name, but use the full user input for the XML search.

CCBUG:299415