svn: Fix regex used for old login scheme detection.
The previous version worked on modules that had *no* user (a setup most
of mine had as the user was set in the ssh config).

For modules with a user set the trailing '@' was included in the uid
which make the check fail even when the module was correct. (Though I
had thought I had tested this case with other modules).